Implemented 2nd render target (normals), implemented SRV for depth buffer

+++ Task6_1_AO/siMesh.h
@@ -19,7 +19,7 @@ class siMesh
 
    void initBuffer(ID3D12Device* device, const siCommandList& commandList);
 public:
-   [[nodiscard]] UINT getIndexCount() const { return indices.size(); }
+   [[nodiscard]] UINT getIndexCount() const { return static_cast<UINT>(indices.size()); }
    [[nodiscard]] const D3D12_VERTEX_BUFFER_VIEW& getVertexBufferView() const { return vertexBufferView; }
    [[nodiscard]] const D3D12_INDEX_BUFFER_VIEW& getIndexBufferView() const { return indexBufferView; }
    [[nodiscard]] const std::string& getDiffuseMap() const { return diffuseMap; }
